chore: updated examples
